🎨 Moves all images and logos into public dir
